+++ N2PW/include/Balance.hpp
@@ -31,9 +31,9 @@ public:
    /* destructor */
    ~Balance();
 
-   void c_unbalance(const int, double *a);
-   void c_balance(const int, double *a);
-   void t_balance(const int, double *a);
+   void c_unbalance(const int, float *a);
+   void c_balance(const int, float *a);
+   void t_balance(const int, float *a);
    void i_balance(const int, int    *a);
 };
