+++ N2PW/include/Pneb.hpp
@@ -12,7 +12,7 @@ class Pneb : public PGrid, public d1db  {
 
    //int ispin,ne[2],neq[2];
    int parallelized;
-   double *s22,*s21,*s12,*s11,*sa1,*sa0,*st1;
+   float *s22,*s21,*s12,*s11,*sa1,*sa0,*st1;
    int *ma[2],*ma1[2],*ma2[2],*mc[2],*na[2],*nc[2];
    int mcq[2],ncq[2];
    int  ncqmax;
@@ -39,24 +39,24 @@ public:
                }
         }
 
-        void g_read(const int, double *);
+        void g_read(const int, float *);
 
 
 
-        double *g_allocate(const int nb) {
-           double *ptr;
-           ptr = new double [2*(neq[0]+neq[1])*npack(nb)];
+        float *g_allocate(const int nb) {
+           float *ptr;
+           ptr = new float [2*(neq[0]+neq[1])*npack(nb)];
            return ptr;
         }
-        void g_deallocate(double *ptr) { delete [] ptr;}
+        void g_deallocate(float *ptr) { delete [] ptr;}
 
 
-        double *h_allocate() {
-           double *ptr;
-           ptr = new double [(neq[0]+neq[1])*n2ft3d];
+        float *h_allocate() {
+           float *ptr;
+           ptr = new float [(neq[0]+neq[1])*n2ft3d];
            return ptr;
         }
-        void h_deallocate(double *ptr) { delete [] ptr;}
+        void h_deallocate(float *ptr) { delete [] ptr;}
 
         int m_size(const int mb) {
            int nsize;
@@ -64,43 +64,43 @@ public:
            else nsize = ne[mb]*ne[mb];
            return nsize;
         }
-        double *m_allocate(const int mb, const int nblock) {
-           double *ptr;
+        float *m_allocate(const int mb, const int nblock) {
+           float *ptr;
            int nsize;
            if (mb==-1)
               nsize = ne[0]*ne[0] + ne[1]*ne[1];
            else
               nsize = ne[mb]*ne[mb];
 
-           ptr = new double [nblock*nsize];
+           ptr = new float [nblock*nsize];
            return ptr;
         }
-        void m_deallocate(double *ptr) { delete [] ptr;}
+        void m_deallocate(float *ptr) { delete [] ptr;}
 
-        double gg_traceall(double *, double *);
-        void gg_copy(double *, double *);
-        void g_zero(double *);
-        void hr_aSumSqr(const double, double *, double *);
-        void ggm_sym_Multiply(double *, double *, double *);
-        void ffm_sym_Multiply(const int, double *, double *, double *);
-        void fmf_Multiply(const int, double *, double *, double, double *, double);
+        float gg_traceall(float *, float *);
+        void gg_copy(float *, float *);
+        void g_zero(float *);
+        void hr_aSumSqr(const float, float *, float *);
+        void ggm_sym_Multiply(float *, float *, float *);
+        void ffm_sym_Multiply(const int, float *, float *, float *);
+        void fmf_Multiply(const int, float *, float *, float, float *, float);
 
-        void m_scal(const double, double *);
-        double m_trace(double *);
-        void m_diagonalize(double *, double *);
-        void mmm_Multiply(const int, double *, double *, double, double*, double);
-        double m_dmax(const int, double *);
-        void m_scale_s22(const int, const double, double *);
-        void m_scale_s21(const int, const double, double *);
-        void m_scale_s11(const int, const double, double *);
+        void m_scal(const float, float *);
+        float m_trace(float *);
+        void m_diagonalize(float *, float *);
+        void mmm_Multiply(const int, float *, float *, float, float*, float);
+        float m_dmax(const int, float *);
+        void m_scale_s22(const int, const float, float *);
+        void m_scale_s21(const int, const float, float *);
+        void m_scale_s11(const int, const float, float *);
 
 
-        void gh_fftb(double *, double *);
-        void ggm_lambda(double, double *, double *, double *);
+        void gh_fftb(float *, float *);
+        void ggm_lambda(float, float *, float *, float *);
 
-        void gg_SMul(double, double *, double *);
-        void gg_Sum2(double *, double *);
-        void ggg_Minus(double *, double *, double *);
+        void gg_SMul(float, float *, float *);
+        void gg_Sum2(float *, float *);
+        void ggg_Minus(float *, float *, float *);
 
 };
