+++ N2PW/include/d3db.hpp
@@ -23,7 +23,7 @@ class d3db : public Mapping3 {
    int ***p_iq_to_i1,***p_iq_to_i2,***p_iz_to_i2;
    int ***p_i1_start,***p_i2_start;
 
-   double *tmpx,*tmpy,*tmpz;
+   float *tmpx,*tmpy,*tmpz;
 
 public:
         Parallel  *parall;
@@ -36,41 +36,41 @@ public:
         ~d3db();
 
         /* r array operators */
-        double * r_alloc();
-        double * r_nalloc(const int);
-        void     r_dealloc(double *);
-        void     r_zero(double *);
-        void     rr_copy(const double *,double *);
-        void     rr_SMul(const double, const double *, double *);
-        void	 r_SMul(const double, double *);
-        double	 r_dsum(const double *);
-        double	 rr_dot(const double *, const double *);
-        void  	 r_zero_ends(double *);
-        void  	 r_abs(double *);
-        void  	 rrr_Sum(const double *, const double *, double *);
-        void  	 rrr_Mul(const double *, const double *, double *);
-        void  	 rrr_Minus(const double *, const double *, double *);
-        void  	 rrr_Divide(const double *, const double *, double *);
-        void  	 rr_daxpy(const double, const double *, double *);
-        //void  	 r_read(const int, const int, double *);
-	void     c_read(const int, double *, const int);
+        float * r_alloc();
+        float * r_nalloc(const int);
+        void     r_dealloc(float *);
+        void     r_zero(float *);
+        void     rr_copy(const float *,float *);
+        void     rr_SMul(const float, const float *, float *);
+        void	 r_SMul(const float, float *);
+        float	 r_dsum(const float *);
+        float	 rr_dot(const float *, const float *);
+        void  	 r_zero_ends(float *);
+        void  	 r_abs(float *);
+        void  	 rrr_Sum(const float *, const float *, float *);
+        void  	 rrr_Mul(const float *, const float *, float *);
+        void  	 rrr_Minus(const float *, const float *, float *);
+        void  	 rrr_Divide(const float *, const float *, float *);
+        void  	 rr_daxpy(const float, const float *, float *);
+        //void  	 r_read(const int, const int, float *);
+	void     c_read(const int, float *, const int);
 
-	void     cr_fft3d(double *);
-	void     rc_fft3d(double *);
+	void     cr_fft3d(float *);
+	void     rc_fft3d(float *);
 
-	void     c_transpose_jk(double *, double *, double *);
-	void     t_transpose_jk(double *, double *, double *);
+	void     c_transpose_jk(float *, float *, float *);
+	void     t_transpose_jk(float *, float *, float *);
 
-	void     c_transpose_ijk(const int, double *, double *, double *);
-	void     t_transpose_ijk(const int, double *, double *, double *);
+	void     c_transpose_ijk(const int, float *, float *, float *);
+	void     t_transpose_ijk(const int, float *, float *, float *);
 
-	void     c_timereverse(double *, double *, double *);
+	void     c_timereverse(float *, float *, float *);
         int	 timereverse_size();
 
         /* t array operators */
-        double * t_alloc();
-        void     t_dealloc(double *);
-	void     t_read(const int, double *, const int);
+        float * t_alloc();
+        void     t_dealloc(float *);
+	void     t_read(const int, float *, const int);
 };
 
 #endif
