+++ N2PW/include/v_exc.hpp
@@ -2,7 +2,7 @@
 #define _VEXC_H_
 
 
-extern void v_exc(const int, const int, double *, double *, double *, double *);
+extern void v_exc(const int, const int, float *, float *, float *, float *);
 
 
 #endif
