+++ N2PW/nwpwlib/D3dB/Pneb.cpp
@@ -35,7 +35,7 @@ Pneb::Pneb(Parallel *inparall, int ispin, int *ne) : PGrid(inparall), d1db(inpar
     int np_j = d1db::parall->np_j();
     parallelized = (np_j>1);
 
-    s22 = new double[7*ne[0]*ne[0]];
+    s22 = new float[7*ne[0]*ne[0]];
     s21 = &s22[1*ne[0]*ne[0]];
     s12 = &s22[2*ne[0]*ne[0]];
     s11 = &s22[3*ne[0]*ne[0]];
@@ -60,10 +60,10 @@ Pneb::Pneb(Parallel *inparall, int ispin, int *ne) : PGrid(inparall), d1db(inpar
     }
 }
 
-void Pneb::g_read(const int iunit, double *psi)
+void Pneb::g_read(const int iunit, float *psi)
 {
    int ms,n,indx,i,pj,qj,taskid_j;
-   double *tmp2 = new double[n2ft3d];
+   float *tmp2 = new float[n2ft3d];
 
    taskid_j = d1db::parall->taskid_j();
    for (ms=0; ms<ispin; ++ms)
@@ -83,10 +83,10 @@ void Pneb::g_read(const int iunit, double *psi)
    delete [] tmp2;
 }
 
-double Pneb::gg_traceall(double *psi1, double *psi2)
+float Pneb::gg_traceall(float *psi1, float *psi2)
 {
    int n,indx;
-   double sum=0.0;
+   float sum=0.0;
 
    indx = 0;
    for (n=0; n<(neq[0]+neq[1]); ++n)
@@ -100,20 +100,20 @@ double Pneb::gg_traceall(double *psi1, double *psi2)
    return sum;
 }
 
-void Pneb::gg_copy(double *psi1, double *psi2)
+void Pneb::gg_copy(float *psi1, float *psi2)
 {
    int one=1;
    int nsize = 2*(neq[0]+neq[1])*npack(1);
    dcopy_(&nsize,psi1,&one,psi2,&one);
 }
-void Pneb::gg_SMul(double alpha,double *psi1, double *psi2)
+void Pneb::gg_SMul(float alpha,float *psi1, float *psi2)
 {
    int i;
    int nsize = 2*(neq[0]+neq[1])*npack(1);
    for (i=0; i<nsize; ++i) 
       psi2[i] = alpha*psi1[i];
 }
-void Pneb::gg_Sum2(double *psi1, double *psi2)
+void Pneb::gg_Sum2(float *psi1, float *psi2)
 {
    int i;
    int nsize = 2*(neq[0]+neq[1])*npack(1);
@@ -121,7 +121,7 @@ void Pneb::gg_Sum2(double *psi1, double *psi2)
       psi2[i] += psi1[i];
 }
 
-void Pneb::ggg_Minus(double *psi1, double *psi2, double *psi3)
+void Pneb::ggg_Minus(float *psi1, float *psi2, float *psi3)
 {
    int i;
    int nsize = 2*(neq[0]+neq[1])*npack(1);
@@ -130,16 +130,16 @@ void Pneb::ggg_Minus(double *psi1, double *psi2, double *psi3)
 }
 
 
-void Pneb::g_zero(double *psi2)
+void Pneb::g_zero(float *psi2)
 {
    int one=1;
    int zero=0;
    int nsize = 2*(neq[0]+neq[1])*npack(1);
-   double rzero=0.0;
+   float rzero=0.0;
   
    dcopy_(&nsize,&rzero,&zero,psi2,&one);
 }
-void Pneb::gh_fftb(double *psi, double *psi_r)
+void Pneb::gh_fftb(float *psi, float *psi_r)
 {
    int n,done;
    int indx1,indx1n,shift1;
@@ -170,13 +170,13 @@ void Pneb::gh_fftb(double *psi, double *psi_r)
 
 }
 
-void Pneb::hr_aSumSqr(const double alpha, double *psir, double *dn)
+void Pneb::hr_aSumSqr(const float alpha, float *psir, float *dn)
 {
    int n,ms,k,indx0,indx1;
    int one=1;
    int zero=0;
    int nsize = n2ft3d*ispin;
-   double rzero = 0.0;
+   float rzero = 0.0;
 
    dcopy_(&nsize,&rzero,&zero,dn,&one);
    indx0 = 0;
@@ -196,7 +196,7 @@ void Pneb::hr_aSumSqr(const double alpha, double *psir, double *dn)
 
 
 
-void Pneb::ggm_sym_Multiply(double *psi1, double *psi2, double *hml)
+void Pneb::ggm_sym_Multiply(float *psi1, float *psi2, float *hml)
 {
    int ms,j,k,n,shift0,shift1,mshift0,mshift1;
 
@@ -204,10 +204,10 @@ void Pneb::ggm_sym_Multiply(double *psi1, double *psi2, double *hml)
    int ng  = 2*npack(1);
    int ng0 = 2*nzero(1);
 
-   double rzero = 0.0;
-   double rtwo  = 2.0;
-   double rone =  1.0;
-   double rmone = -1.0;
+   float rzero = 0.0;
+   float rtwo  = 2.0;
+   float rone =  1.0;
+   float rmone = -1.0;
 
    if (parallelized)
    {
@@ -253,7 +253,7 @@ void Pneb::ggm_sym_Multiply(double *psi1, double *psi2, double *hml)
 
 
 
-void Pneb::ffm_sym_Multiply(const int mb, double *psi1, double *psi2, double *hml)
+void Pneb::ffm_sym_Multiply(const int mb, float *psi1, float *psi2, float *hml)
 {
    int ms,ms1,ms2,ishift2,j,k,n,shift0,shift1,mshift0,mshift1,nn;
 
@@ -261,10 +261,10 @@ void Pneb::ffm_sym_Multiply(const int mb, double *psi1, double *psi2, double *hm
    int ng  = 2*npack(1);
    int ng0 = 2*nzero(1);
 
-   double rzero = 0.0;
-   double rtwo  = 2.0;
-   double rone =  1.0;
-   double rmone = -1.0;
+   float rzero = 0.0;
+   float rtwo  = 2.0;
+   float rone =  1.0;
+   float rmone = -1.0;
 
    if (parallelized)
    {
@@ -319,7 +319,7 @@ void Pneb::ffm_sym_Multiply(const int mb, double *psi1, double *psi2, double *hm
 
 
 
-void Pneb::fmf_Multiply(const int mb, double *psi1, double *hml, double alpha, double *psi2, double beta)
+void Pneb::fmf_Multiply(const int mb, float *psi1, float *hml, float alpha, float *psi2, float beta)
 {
    int ms,ms1,ms2,n,shift1,mshift1,ishift2;
    int ng  = 2*npack(1);
@@ -358,7 +358,7 @@ void Pneb::fmf_Multiply(const int mb, double *psi1, double *hml, double alpha, d
 
 
 
-double Pneb::m_dmax(const int mb, double *hml)
+float Pneb::m_dmax(const int mb, float *hml)
 {
   int one = 1;
   int nn;
@@ -372,18 +372,18 @@ double Pneb::m_dmax(const int mb, double *hml)
 
 
 
-void Pneb::m_scal(double alpha, double *hml)
+void Pneb::m_scal(float alpha, float *hml)
 {
   int one = 1;
   int nsize = ne[0]*ne[0] + ne[1]*ne[1];
   dscal_(&nsize,&alpha,hml,&one);
 }
 
-double Pneb::m_trace(double *hml)
+float Pneb::m_trace(float *hml)
 {
    int ms,i;
    int mshift = 0;
-   double sum = 0.0;
+   float sum = 0.0;
    for (ms=0; ms<ispin; ++ms)
    {
       for (i=0; i<ne[ms]; ++i)
@@ -393,11 +393,11 @@ double Pneb::m_trace(double *hml)
    return sum;
 }
 
-void Pneb::m_diagonalize(double *hml, double *eig)
+void Pneb::m_diagonalize(float *hml, float *eig)
 {
    int ms,nn,shift1,shift2;
    int n,ierr;
-   double *xmp1;
+   float *xmp1;
    if (parallelized)
    {
      printf("not finished\n");
@@ -405,7 +405,7 @@ void Pneb::m_diagonalize(double *hml, double *eig)
    else
    {
       nn  = 2*ne[0]*ne[0];
-      xmp1 = new double[nn];
+      xmp1 = new float[nn];
       shift1 = 0;
       shift2 = 0;
       for (ms=0; ms<ispin; ++ms)
@@ -426,7 +426,7 @@ void Pneb::m_diagonalize(double *hml, double *eig)
 }
 
 
-void Pneb::m_scale_s22(const int mb, const double dte, double *s22)
+void Pneb::m_scale_s22(const int mb, const float dte, float *s22)
 {
    int j,k,ms,ms1,ms2,ishift2,indx0,indx,indxt;
 
@@ -455,7 +455,7 @@ void Pneb::m_scale_s22(const int mb, const double dte, double *s22)
    }
 }
 
-void Pneb::m_scale_s21(const int mb, const double dte, double *s21)
+void Pneb::m_scale_s21(const int mb, const float dte, float *s21)
 {
    int j,k,ms,ms1,ms2,ishift2,indx0,indx,indxt;
    if (mb==-1)
@@ -484,7 +484,7 @@ void Pneb::m_scale_s21(const int mb, const double dte, double *s21)
    }
 }
 
-void Pneb::mmm_Multiply(const int mb, double *a, double *b, double alpha, double *c, double beta)
+void Pneb::mmm_Multiply(const int mb, float *a, float *b, float alpha, float *c, float beta)
 {
    int ms,n,ms1,ms2,ishift2,shift2;
    if (mb==-1)
@@ -508,7 +508,7 @@ void Pneb::mmm_Multiply(const int mb, double *a, double *b, double alpha, double
 }
 
 
-void Pneb::m_scale_s11(const int mb, const double dte, double *s11)
+void Pneb::m_scale_s11(const int mb, const float dte, float *s11)
 {
    int j,k,ms,ms1,ms2,ishift2,indx0,indx,indxt;
    if (mb==-1)
@@ -544,12 +544,12 @@ void Pneb::m_scale_s11(const int mb, const double dte, double *s11)
 #define CONVGLMD        1e-15
 
 
-void Pneb::ggm_lambda(double dte,double *psi1, double *psi2, double *lmbda)
+void Pneb::ggm_lambda(float dte,float *psi1, float *psi2, float *lmbda)
 {
    int one=1;
    int ms,nn,ii,done;
-   double adiff;
-   double rmone = -1.0;
+   float adiff;
+   float rmone = -1.0;
 
    for (ms=0; ms<ispin; ++ms)
    {
